+++ src/Widget.php
@@ -81,7 +81,7 @@ abstract class Widget implements NoEncodeStringableInterface
      * @param array $config The configuration for creating a widget. For a description of the configuration syntax, see
      * array definitions documentation in the Yii Definitions by link
      * {@link https://github.com/yiisoft/definitions#arraydefinition).
-     * @param string|null $theme Widget theme.
+     * @param string|null $theme The widget theme.
      *
      * @throws InvalidConfigException
      * @throws CircularReferenceException
@@ -96,17 +96,25 @@ abstract class Widget implements NoEncodeStringableInterface
         ?string $theme = null
     ): static {
         $config = ArrayDefinitionHelper::merge(
-            static::getDefaultConfig($theme),
+            static::getThemeConfig($theme),
             $config,
             empty($constructorArguments) ? [] : [ArrayDefinition::CONSTRUCTOR => $constructorArguments],
         );
 
         $config[ArrayDefinition::CLASS_NAME] = static::class;
 
-        return WidgetFactory::createWidget($config);
+        return WidgetFactory::createWidget($config, $theme);
     }
 
-    protected static function getDefaultConfig(?string $theme): array
+    /**
+     * Returns configuration that will be merged with configuration passed to {@see widget()} method.
+     *
+     * @param string|null $theme The widget theme.
+     *
+     * @return array Configuration in the form of array definition (see syntax description in the Yii Definitions
+     * documentation by link {@link https://github.com/yiisoft/definitions#arraydefinition}).
+     */
+    protected static function getThemeConfig(?string $theme): array
     {
         return [];
     }