+++ src/WidgetFactory.php
@@ -4,10 +4,13 @@ declare(strict_types=1);
 
 namespace Yiisoft\Widget;
 
+use InvalidArgumentException;
 use Psr\Container\ContainerInterface;
+use Yiisoft\Definitions\ArrayDefinition;
 use Yiisoft\Definitions\Exception\CircularReferenceException;
 use Yiisoft\Definitions\Exception\InvalidConfigException;
 use Yiisoft\Definitions\Exception\NotInstantiableException;
+use Yiisoft\Definitions\Helpers\ArrayDefinitionHelper;
 use Yiisoft\Factory\NotFoundException;
 use Yiisoft\Factory\Factory;
 
@@ -20,12 +23,18 @@ final class WidgetFactory
 {
     private static ?Factory $factory = null;
 
+    /**
+     * @psalm-var array<string, array<string, array>>
+     */
+    private static array $themes = [];
+
     private function __construct()
     {
     }
 
     /**
      * @psalm-param array<string, mixed> $definitions
+     * @psalm-param array<string, array<string, array>> $themes
      *
      * @throws InvalidConfigException
      *
@@ -34,9 +43,11 @@ final class WidgetFactory
     public static function initialize(
         ContainerInterface $container,
         array $definitions = [],
+        array $themes = [],
         bool $validate = true
     ): void {
         self::$factory = new Factory($container, $definitions, $validate);
+        self::$themes = $themes;
     }
 
     /**
@@ -55,12 +66,31 @@ final class WidgetFactory
      * @psalm-suppress MixedInferredReturnType
      * @psalm-suppress MixedReturnStatement
      */
-    public static function createWidget(array $config): Widget
+    public static function createWidget(array $config, ?string $theme): Widget
     {
         if (self::$factory === null) {
             throw new WidgetFactoryInitializationException(
                 'Widget factory should be initialized with WidgetFactory::initialize() call.',
             );
+        }
+
+        if ($theme !== null) {
+            if (!isset(self::$themes[$theme])) {
+                throw new InvalidArgumentException(sprintf('Theme "%s" not found.', $theme));
+            }
+
+            if (
+                is_string($config[ArrayDefinition::CLASS_NAME])
+                && isset(
+                    $config[ArrayDefinition::CLASS_NAME],
+                    self::$themes[$theme][$config[ArrayDefinition::CLASS_NAME]]
+                )
+            ) {
+                $config = ArrayDefinitionHelper::merge(
+                    self::$themes[$theme][$config[ArrayDefinition::CLASS_NAME]],
+                    $config
+                );
+            }
         }
 
         return self::$factory->create($config);